+++ spec/strategies/by_url_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe Faraday::HttpCache::Storage do
+describe Faraday::HttpCache::Strategies::ByUrl do
   let(:cache_key) { '6e3b941d0f7572291c777b3e48c04b74124a55d0' }
   let(:request) do
     env = { method: :get, url: 'http://test/index' }
@@ -13,41 +13,26 @@ describe Faraday::HttpCache::Storage do
 
   let(:cache) { Faraday::HttpCache::MemoryStore.new }
 
-  let(:storage) { Faraday::HttpCache::Storage.new(store: cache) }
-  subject { storage }
-
-  before do
-    allow(Kernel).to receive(:warn).with(
-      'Deprecated: Faraday::HttpCache::Storage is deprecated and will be removed '\
-      'in the next major release. Use Faraday::HttpCache::Strategies::ByUrl instead.'
-    )
-  end
-
-  it 'creates strategy and warns about deprecation' do
-    expect(Kernel).to receive(:warn).with(
-      'Deprecated: Faraday::HttpCache::Storage is deprecated and will be removed '\
-      'in the next major release. Use Faraday::HttpCache::Strategies::ByUrl instead.'
-    )
-    is_expected.to be_a_kind_of(Faraday::HttpCache::Strategies::ByUrl)
-  end
+  let(:strategy) { described_class.new(store: cache) }
+  subject { strategy }
 
   describe 'Cache configuration' do
     it 'uses a MemoryStore by default' do
       expect(Faraday::HttpCache::MemoryStore).to receive(:new).and_call_original
-      Faraday::HttpCache::Storage.new
+      described_class.new
     end
 
     it 'raises an error when the given store is not valid' do
       wrong = double
 
       expect {
-        Faraday::HttpCache::Storage.new(store: wrong)
+        described_class.new(store: wrong)
       }.to raise_error(ArgumentError)
     end
   end
 
   describe 'storing responses' do
-    shared_examples 'A storage with serialization' do
+    shared_examples 'A strategy with serialization' do
       it 'writes the response object to the underlying cache' do
         entry = [serializer.dump(request.serializable_hash), serializer.dump(response.serializable_hash)]
         expect(cache).to receive(:write).with(cache_key, [entry])
@@ -57,7 +42,7 @@ describe Faraday::HttpCache::Storage do
 
     context 'with the JSON serializer' do
       let(:serializer) { JSON }
-      it_behaves_like 'A storage with serialization'
+      it_behaves_like 'A strategy with serialization'
 
       context 'when ASCII characters in response cannot be converted to UTF-8', if: Gem::Version.new(RUBY_VERSION) < Gem::Version.new('3.1') do
         let(:response) do
@@ -67,10 +52,10 @@ describe Faraday::HttpCache::Storage do
 
         it 'raises and logs a warning' do
           logger = double(:logger, warn: nil)
-          storage = Faraday::HttpCache::Storage.new(logger: logger)
+          strategy = described_class.new(logger: logger)
 
           expect {
-            storage.write(request, response)
+            strategy.write(request, response)
           }.to raise_error(::Encoding::UndefinedConversionError)
           expect(logger).to have_received(:warn).with(
             'Response could not be serialized: "\xE2" from ASCII-8BIT to UTF-8. Try using Marshal to serialize.'
@@ -82,16 +67,16 @@ describe Faraday::HttpCache::Storage do
     context 'with the Marshal serializer' do
       let(:cache_key) { '337d1e9c6c92423dd1c48a23054139058f97be40' }
       let(:serializer) { Marshal }
-      let(:storage) { Faraday::HttpCache::Storage.new(store: cache, serializer: Marshal) }
+      let(:strategy) { described_class.new(store: cache, serializer: Marshal) }
 
-      it_behaves_like 'A storage with serialization'
+      it_behaves_like 'A strategy with serialization'
     end
   end
 
   describe 'reading responses' do
-    let(:storage) { Faraday::HttpCache::Storage.new(store: cache, serializer: serializer) }
+    let(:strategy) { described_class.new(store: cache, serializer: serializer) }
 
-    shared_examples 'A storage with serialization' do
+    shared_examples 'A strategy with serialization' do
       it 'returns nil if the response is not cached' do
         expect(subject.read(request)).to be_nil
       end
@@ -106,13 +91,13 @@ describe Faraday::HttpCache::Storage do
     context 'with the JSON serializer' do
       let(:serializer) { JSON }
 
-      it_behaves_like 'A storage with serialization'
+      it_behaves_like 'A strategy with serialization'
     end
 
     context 'with the Marshal serializer' do
       let(:serializer) { Marshal }
 
-      it_behaves_like 'A storage with serialization'
+      it_behaves_like 'A strategy with serialization'
     end
   end
 
@@ -143,10 +128,9 @@ describe Faraday::HttpCache::Storage do
     end
 
     it 'is fresh until cached and that 1 second elapses then the response is no longer fresh' do
-      current_time = Time.now
       headers = {
-        'Date' => (current_time - 39).httpdate,
-        'Expires' => (current_time + 40).httpdate
+        'Date' => (Time.now - 39).httpdate,
+        'Expires' => (Time.now + 40).httpdate
       }
 
       response = Faraday::HttpCache::Response.new(response_headers: headers)